requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
gradio>=3.40.0
streamlit>=1.25.0
huggingface-hub>=0.16.0
safetensors>=0.3.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
tqdm>=4.65.0
matplotlib>=3.7.0